+++ src/duplocloud/args.py
@@ -36,7 +36,7 @@ TENANT = Arg("tenant", "-T",
              help='The tenant name',
              default=os.getenv('DUPLO_TENANT', None))
 
-TENANT_ID = Arg("tenant-id", "--tenantid", "--tid",
+TENANT_ID = Arg("tenantid", "--tenant-id", "--tid",
              help='The tenant id',
              default=os.getenv('DUPLO_TENANT_ID', None))