+++ src/duplocloud/client.py
@@ -40,7 +40,7 @@ class DuploClient():
                host: args.HOST=None,
                token: args.TOKEN=None,
                tenant: args.TENANT=None,
-               tenantid: args.TENANT_ID=None,
+               tenant_id: args.TENANT_ID=None,
                home_dir: args.HOME_DIR = None,
                config_file: args.CONFIG = None,
                cache_dir: args.CACHE_DIR = None,
@@ -64,6 +64,9 @@ class DuploClient():
     # ignore the given token with interactive mode
     if token and interactive: 
       token = None
+    # if a tenant id was given, the tenant name must be ignored
+    if tenant_id:
+      tenant = None
 
     user_home = Path.home()
     self.home_dir = home_dir or f"{user_home}/.duplo"
@@ -74,7 +77,7 @@ class DuploClient():
     self.__host = self.__sanitize_host(host)
     self.__token = token.strip() if token else token
     self.__tenant = tenant.strip().lower() if tenant else tenant
-    self.tenantid = tenantid
+    self.tenantid = tenant_id.strip() if tenant_id else tenant_id
     self.version = version
     self.interactive = interactive
     self.nocache = nocache